remove -Wno-error=sign-compare, make warning as error (#21358)
* remove -Wno-error=sign-compare, make warning as error

test=develop test=document_fix

* fix exist compile warning

test=develop

+++ paddle/fluid/operators/concat_op.h
@@ -26,7 +26,7 @@ namespace paddle {
 namespace operators {
 static inline framework::DDim ComputeAndCheckShape(
     const bool is_runtime, const std::vector<framework::DDim>& inputs_dims,
-    const int axis) {
+    const size_t axis) {
   const size_t n = inputs_dims.size();
   auto out_dims = inputs_dims[0];
   size_t in_zero_dims_size = out_dims.size();
